Revise for clarify the message (Persistent volume)
I know the persistent volume mean. However, people, who is not familiar with the k8s, might not understand it. I guess you want to "expore" the documentthat is why you don't clearly mentioned that is environment valiables. To avoid the confusion I adject the term "persistent volume" into secrets as file which is mentioned on the document